requirements.txt
requests
pytest
pytest-html
pysmb
samba
Appium-Python-Client
selenium==4.8.0
pytest-xdist>=3.2.0
requests>=2.28.0
urllib3>=1.26.0